make server push packet as abstract class
