+++ brpc-java-core/src/main/java/com/baidu/brpc/client/BrpcProxy.java
@@ -38,7 +38,7 @@ import com.baidu.brpc.protocol.Response;
 import com.baidu.brpc.protocol.nshead.NSHead;
 import com.baidu.brpc.protocol.nshead.NSHeadMeta;
 import com.baidu.brpc.protocol.push.SPHead;
-import com.baidu.brpc.protocol.push.base.ServerPushProtocol;
+import com.baidu.brpc.protocol.push.ServerPushProtocol;
 import com.baidu.brpc.utils.ProtobufUtils;
 
 import lombok.extern.slf4j.Slf4j;
