--- brpc-java-core/src/main/java/com/baidu/brpc/protocol/push/RegistryContent.java
@@ -1,15 +1,0 @@
-package com.baidu.brpc.protocol.push;
-
-public class RegistryContent {
-
-    private String clientName;
-
-    public String getClientName() {
-        return clientName;
-    }
-
-    public void setClientName(String clientName) {
-        this.clientName = clientName;
-    }
-
-}
